Child Details added and QR implemented

+++ app/src/main/java/com/example/youthconnect/SignUp.kt
@@ -306,7 +306,7 @@ fun FirstSignupFormScreen() {
             val parentsID : List<String>  = listOf(parentID)
             val parent = Parent(parentFullName,
                 parentID,
-                parentPhoneNumber.toInt(),
+                parentPhoneNumber,
                 password)
 
             val child = Child(childFullName,
